elastic and hyperelastic distinction
