+++ src/constitutive/hyperelastic/mod.rs
@@ -1,7 +1,14 @@
+//! Hyperelastic constitutive models.
+//!
+//! Hyperelastic constitutive models are derived from and defined by a helmholtz free energy density. As a result, the tangent stiffness associated with the first Piola-Kirchoff stress is symmetric for hyperelastic constitutive models.
+//!
+//! ```math
+//! \mathcal{C}_{iJkL} = \mathcal{C}_{kLiJ}
+//! ```
+
 #[cfg(test)]
 mod test;
 
-mod almansi_hamel;
 mod additive_decomposition;
 mod arruda_boyce;
 mod gent;
@@ -11,7 +18,6 @@ mod yeoh;
 
 pub use self::
 {
-    almansi_hamel::AlmansiHamelModel,
     arruda_boyce::ArrudaBoyceModel,
     gent::GentModel,
     mooney_rivlin::MooneyRivlinModel,
@@ -23,6 +29,12 @@ use super::*;
 /// Required methods for hyperelastic constitutive models.
 pub trait HyperelasticConstitutiveModel
 {
+    /// Calculates and returns the Helmholtz free energy density.
+    ///
+    /// ```math
+    /// a = a(\mathbf{F})
+    /// ```
+    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar;
     /// Returns the bulk modulus.
     fn get_bulk_modulus(&self) -> &Scalar;
     /// Returns the shear modulus.
