+++ src/constitutive/hyperelastic/neo_hookean/mod.rs
@@ -44,16 +44,6 @@ impl<'a> ConstitutiveModel<'a> for NeoHookeanModel<'a>
         let scaled_shear_modulus = self.get_shear_modulus()/jacobian.powf(5.0/3.0);
         (CauchyTangentStiffness::dyad_ik_jl(&identity, deformation_gradient) + CauchyTangentStiffness::dyad_il_jk(deformation_gradient, &identity) - CauchyTangentStiffness::dyad_ij_kl(&identity, deformation_gradient)*(2.0/3.0))*scaled_shear_modulus + CauchyTangentStiffness::dyad_ij_kl(&(identity*(0.5*self.get_bulk_modulus()*(jacobian + 1.0/jacobian)) - self.calculate_left_cauchy_green_deformation(deformation_gradient).deviatoric()*(scaled_shear_modulus*5.0/3.0)), &inverse_transpose_deformation_gradient)
     }
-    /// Calculates and returns the Helmholtz free energy density.
-    ///
-    /// ```math
-    /// a(\mathbf{F}) = \frac{\mu}{2}\left[\mathrm{tr}(\mathbf{B}^*) - 3\right] + \frac{\kappa}{2}\left[\frac{1}{2}\left(J^2 - 1\right) - \ln J\right]
-    /// ```
-    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar
-    {
-        let jacobian = deformation_gradient.determinant();
-        0.5*(self.get_shear_modulus()*(self.calculate_left_cauchy_green_deformation(deformation_gradient).trace()/jacobian.powf(2.0/3.0) - 3.0) + self.get_bulk_modulus()*(0.5*(jacobian.powi(2) - 1.0) - jacobian.ln()))
-    }
     fn new(parameters: ConstitutiveModelParameters<'a>) -> Self
     {
         Self
@@ -66,6 +56,16 @@ impl<'a> ConstitutiveModel<'a> for NeoHookeanModel<'a>
 /// Hyperelastic constitutive model implementation of the Neo-Hookean hyperelastic constitutive model.
 impl<'a> HyperelasticConstitutiveModel for NeoHookeanModel<'a>
 {
+    /// Calculates and returns the Helmholtz free energy density.
+    ///
+    /// ```math
+    /// a(\mathbf{F}) = \frac{\mu}{2}\left[\mathrm{tr}(\mathbf{B}^*) - 3\right] + \frac{\kappa}{2}\left[\frac{1}{2}\left(J^2 - 1\right) - \ln J\right]
+    /// ```
+    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar
+    {
+        let jacobian = deformation_gradient.determinant();
+        0.5*(self.get_shear_modulus()*(self.calculate_left_cauchy_green_deformation(deformation_gradient).trace()/jacobian.powf(2.0/3.0) - 3.0) + self.get_bulk_modulus()*(0.5*(jacobian.powi(2) - 1.0) - jacobian.ln()))
+    }
     fn get_bulk_modulus(&self) -> &Scalar
     {
         &self.parameters[0]
