+++ src/constitutive/hyperelastic/yeoh/mod.rs
@@ -69,17 +69,6 @@ impl<'a> ConstitutiveModel<'a> for YeohModel<'a>
         let last_term = CauchyTangentStiffness::dyad_ij_kl(&deviatoric_left_cauchy_green_deformation, &((left_cauchy_green_deformation.deviatoric() * &inverse_transpose_deformation_gradient) * (2.0*self.get_extra_moduli().iter().enumerate().map(|(n, modulus)| ((n as Scalar) + 2.0)*((n as Scalar) + 1.0)*modulus*scalar_term.powi(n.try_into().unwrap())).sum::<Scalar>()/jacobian.powf(7.0/3.0))));
         (CauchyTangentStiffness::dyad_ik_jl(&identity, deformation_gradient) + CauchyTangentStiffness::dyad_il_jk(deformation_gradient, &identity) - CauchyTangentStiffness::dyad_ij_kl(&identity, deformation_gradient)*(2.0/3.0))*scaled_modulus + CauchyTangentStiffness::dyad_ij_kl(&(identity*(0.5*self.get_bulk_modulus()*(jacobian + 1.0/jacobian)) - deviatoric_left_cauchy_green_deformation*(scaled_modulus*5.0/3.0)), &inverse_transpose_deformation_gradient) + last_term
     }
-    /// Calculates and returns the Helmholtz free energy density.
-    ///
-    /// ```math
-    /// a(\mathbf{F}) = \sum_{n=1}^N \frac{\mu_n}{2}\left[\mathrm{tr}(\mathbf{B}^* ) - 3\right]^n + \frac{\kappa}{2}\left[\frac{1}{2}\left(J^2 - 1\right) - \ln J\right]
-    /// ```
-    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar
-    {
-        let jacobian = deformation_gradient.determinant();
-        let scalar_term = self.calculate_left_cauchy_green_deformation(deformation_gradient).trace()/jacobian.powf(2.0/3.0) - 3.0;
-        0.5*(self.get_moduli().iter().enumerate().map(|(n, modulus)| modulus*scalar_term.powi(<usize as TryInto<i32>>::try_into(n).unwrap() + 1)).sum::<Scalar>() + self.get_bulk_modulus()*(0.5*(jacobian.powi(2) - 1.0) - jacobian.ln()))
-    }
     fn new(parameters: ConstitutiveModelParameters<'a>) -> Self
     {
         Self
@@ -92,6 +81,17 @@ impl<'a> ConstitutiveModel<'a> for YeohModel<'a>
 /// Hyperelastic constitutive model implementation of the Yeoh hyperelastic constitutive model.
 impl<'a> HyperelasticConstitutiveModel for YeohModel<'a>
 {
+    /// Calculates and returns the Helmholtz free energy density.
+    ///
+    /// ```math
+    /// a(\mathbf{F}) = \sum_{n=1}^N \frac{\mu_n}{2}\left[\mathrm{tr}(\mathbf{B}^* ) - 3\right]^n + \frac{\kappa}{2}\left[\frac{1}{2}\left(J^2 - 1\right) - \ln J\right]
+    /// ```
+    fn calculate_helmholtz_free_energy_density(&self, deformation_gradient: &DeformationGradient) -> Scalar
+    {
+        let jacobian = deformation_gradient.determinant();
+        let scalar_term = self.calculate_left_cauchy_green_deformation(deformation_gradient).trace()/jacobian.powf(2.0/3.0) - 3.0;
+        0.5*(self.get_moduli().iter().enumerate().map(|(n, modulus)| modulus*scalar_term.powi(<usize as TryInto<i32>>::try_into(n).unwrap() + 1)).sum::<Scalar>() + self.get_bulk_modulus()*(0.5*(jacobian.powi(2) - 1.0) - jacobian.ln()))
+    }
     fn get_bulk_modulus(&self) -> &Scalar
     {
         &self.parameters[0]
